requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
defusedxml==0.5.0
Django==2.0.2
django-braces==1.12.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
ecdsa==0.13
flake8==3.5.0
future==0.16.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
oauth2-provider==0.0
oauthlib==2.0.6
openapi-codec==1.3.2
pew==1.1.2
pipenv==9.0.3
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.4.9
pyflakes==1.6.0
PyJWT==1.5.3
python-jose==2.0.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
